+++ lib/Listeners/FileChanged.php
@@ -42,7 +42,7 @@ use OCP\FullTextSearch\Model\IIndex;
 /**
  * Class FileChanged
  *
- * @package OCA\Circles\Events
+ * @package OCA\Files_FullTextSearch\Listeners
  */
 class FileChanged extends ListenersCore implements IEventListener {
 
@@ -51,7 +51,7 @@ class FileChanged extends ListenersCore implements IEventListener {
 	 * @param Event $event
 	 */
 	public function handle(Event $event): void {
-		if (!($event instanceof NodeWrittenEvent)) {
+		if (!$this->registerFullTextSearchServices() || !($event instanceof NodeWrittenEvent)) {
 			return;
 		}